requirements.txt
python-dotenv
flask
pydantic
flask_cors
